fixed issue not properly logging in

+++ routes/web.php
@@ -44,12 +44,12 @@ Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesControl
 
 
 //Admin Food Items
-Route::get('/admin/food-items', 'admin\FoodItemsController@index');
-Route::get('/admin/food-items/create', 'admin\FoodItemsController@create');
-Route::post('/admin/food-items', 'admin\FoodItemsController@store');
-Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit');
-Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update');
-Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete');
+Route::get('/admin/food-items', 'admin\FoodItemsController@index')->middleware('role:Admin');
+Route::get('/admin/food-items/create', 'admin\FoodItemsController@create')->middleware('role:Admin');
+Route::post('/admin/food-items', 'admin\FoodItemsController@store')->middleware('role:Admin');
+Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit')->middleware('role:Admin');
+Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update')->middleware('role:Admin');
+Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete')->middleware('role:Admin');
 
 
 //Admin Customers
@@ -70,22 +70,22 @@ Route::put('/admin/users/{id}', 'admin\UsersController@update')->middleware('rol
 Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete')->middleware('role:Admin');
 
 //Admin Settings
-Route::get('/admin/settings/general', 'admin\SettingController@general');
-Route::put('/admin/settings/general', 'admin\SettingController@saveGeneral');
-Route::get('/admin/settings/seo', 'admin\SettingController@seo');
-Route::put('/admin/settings/seo', 'admin\SettingController@saveSeo');
-Route::get('/admin/settings/social', 'admin\SettingController@social');
-Route::put('/admin/settings/social', 'admin\SettingController@saveSocial');
+Route::get('/admin/settings/general', 'admin\SettingController@general')->middleware('role:Admin');
+Route::put('/admin/settings/general', 'admin\SettingController@saveGeneral')->middleware('role:Admin');
+Route::get('/admin/settings/seo', 'admin\SettingController@seo')->middleware('role:Admin');
+Route::put('/admin/settings/seo', 'admin\SettingController@saveSeo')->middleware('role:Admin');
+Route::get('/admin/settings/social', 'admin\SettingController@social')->middleware('role:Admin');
+Route::put('/admin/settings/social', 'admin\SettingController@saveSocial')->middleware('role:Admin');
 
 
 
 
 //Admin Authentication
 Route::get('/admin/register', function () {
-    return view('admin/register');
+    return view('/admin/register');
 });
 Route::get('/admin/login', function () {
-    return view('admin/login');
+    return view('/admin/login');
 });
 
 
@@ -96,7 +96,7 @@ Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
 
-View::composer(['home', 'pages/about', 'pages/contact', 'pages/offers', 'pages/reservations', 'pages/thank-you', 'menu/all-categories', 'menu/single-menu', ], function ($view) {
+View::composer(['home', 'pages/about', 'pages/contact', 'pages/offers', 'pages/reservations', 'pages/thank-you', 'menu/all-categories', 'menu/single-menu', 'includes/side-menu' ], function ($view) {
     $generalSettings = GeneralSetting::find(1);
     $socialSettings = SocialSetting::find(1);
     $seoSettings = SeoSetting::find(1);
@@ -109,13 +109,13 @@ View::composer(['home', 'pages/about', 'pages/contact', 'pages/offers', 'pages/r
 });
 
 
-Auth::routes();
+// Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+// Route::get('/home', 'HomeController@index')->name('home');
 
-Auth::routes();
+// Auth::routes();
 
-Route::get('/home', 'HomeController@index')->name('home');
+// Route::get('/home', 'HomeController@index')->name('home');
